Declutter source code for Birch
