--- eval.py
@@ -1,262 +1,0 @@
-import sys
-import numpy as np
-import subprocess
-import shlex
-from scipy.stats import pearsonr, spearmanr
-
-
-def evaluate_classification(prediction_index_list, labels):
-    acc = get_acc(prediction_index_list, labels)
-    pre, rec, f1 = get_pre_rec_f1(prediction_index_list, labels)
-    return acc, pre, rec, f1
-
-
-def evaluate_glue(predictions, true_labels):
-    pearson_r = pearsonr(predictions, true_labels)[0]
-    spearman_r = spearmanr(predictions, true_labels)[0]
-    return pearson_r, spearman_r
-
-
-def evaluate_ner(prediction_index_list, labels, label_map):
-    prediction_index_list, labels = recover_label(prediction_index_list, labels, label_map)
-    assert len(prediction_index_list) == len(labels)
-    print("evaluating {} sentences".format(len(prediction_index_list)))
-    return get_ner_fmeasure(prediction_index_list, labels)
-
-
-def evaluate_trec(predictions_file, qrels_file):
-    cmd = "../Anserini/eval/trec_eval.9.0.4/trec_eval {judgement} {output} -m map -m recip_rank -m P.30".format(
-        judgement=qrels_file, output=predictions_file)
-    pargs = shlex.split(cmd)
-    print("running {}".format(cmd))
-    p = subprocess.Popen(pargs, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    pout, perr = p.communicate()
-
-    if sys.version_info[0] < 3:
-        lines = pout.split(b'\n')
-    else:
-        lines = pout.split(b'\n')
-    map = float(lines[0].strip().split()[-1])
-    mrr = float(lines[1].strip().split()[-1])
-    p30 = float(lines[2].strip().split()[-1])
-    return map, mrr, p30
-
-
-def recover_label(prediction_index_list, labels, label_map):
-    label_reverse_map = {}
-    for k in label_map:
-        label_reverse_map[label_map[k]] = k
-
-    pr, lr = [], []
-    assert len(prediction_index_list) == len(labels)
-    for ps, ls in zip(prediction_index_list, labels):
-        assert len(ps) == len(ls)
-        pr.append([label_reverse_map[t] for t in ps])
-        lr.append([label_reverse_map[t] for t in ls])
-
-    return pr, lr
-
-
-def get_acc(prediction_index_list, labels):
-    acc = sum(np.array(prediction_index_list) == np.array(labels))
-    return acc / (len(labels) + 1e-9)
-
-
-def get_pre_rec_f1(prediction_index_list, labels):
-    tp, tn, fp, fn = 0, 0, 0, 0
-    # print("prediction_index_list: ", prediction_index_list)
-    # print("labels: ", labels)
-    assert len(prediction_index_list) == len(labels)
-    for p, l in zip(prediction_index_list, labels):
-        if p == l:
-            if p == 1:
-                tp += 1
-            else:
-                tn += 1
-        else:
-            if p == 1:
-                fp += 1
-            else:
-                fn += 1
-    eps = 1e-8
-    precision = tp * 1.0 / (tp + fp + eps)
-    recall = tp * 1.0 / (tp + fn + eps)
-    f1 = 2 * precision * recall / (precision + recall + eps)
-    return precision, recall, f1
-
-
-def get_p1(prediction_score_list, labels, data_path, data_name, split):
-    f = open(os.path.join(data_path, "{}/{}_{}.csv".format(data_name, data_name, split)))
-    a2score_label = {}
-    for line, p, l in zip(f, prediction_score_list, labels):
-        label, a, b = line.replace("\n", "").split("\t")
-        if a not in a2score_label:
-            a2score_label[a] = []
-        a2score_label[a].append((p, l))
-
-    acc = 0
-    no_true = 0
-    for a in a2score_label:
-        a2score_label[a] = sorted(a2score_label[a], key=lambda x: x[0], reverse=True)
-        if a2score_label[a][0][1] > 0:
-            acc += 1
-        if sum([tmp[1] for tmp in a2score_label[a]]) == 0:
-            no_true += 1
-
-    p1 = acc / (len(a2score_label) - no_true)
-
-    return p1
-
-
-## input as sentence level labels
-def get_ner_fmeasure(predict_lists, golden_lists, label_type="BMES"):
-    sent_num = len(golden_lists)
-    golden_full = []
-    predict_full = []
-    right_full = []
-    right_tag = 0
-    all_tag = 0
-    for idx in range(0, sent_num):
-        # word_list = sentence_lists[idx]
-        golden_list = golden_lists[idx]
-        predict_list = predict_lists[idx]
-        for idy in range(len(golden_list)):
-            if golden_list[idy] == predict_list[idy]:
-                right_tag += 1
-        all_tag += len(golden_list)
-        if label_type == "BMES" or label_type == "BIOES":
-            gold_matrix = get_ner_BMES(golden_list)
-            pred_matrix = get_ner_BMES(predict_list)
-        else:
-            gold_matrix = get_ner_BIO(golden_list)
-            pred_matrix = get_ner_BIO(predict_list)
-        # print "gold", gold_matrix
-        # print "pred", pred_matrix
-        right_ner = list(set(gold_matrix).intersection(set(pred_matrix)))
-        golden_full += gold_matrix
-        predict_full += pred_matrix
-        right_full += right_ner
-    right_num = len(right_full)
-    golden_num = len(golden_full)
-    predict_num = len(predict_full)
-    if predict_num == 0:
-        precision = -1
-    else:
-        precision = (right_num + 0.0) / predict_num
-    if golden_num == 0:
-        recall = -1
-    else:
-        recall = (right_num + 0.0) / golden_num
-    if (precision == -1) or (recall == -1) or (precision + recall) <= 0.:
-        f_measure = -1
-    else:
-        f_measure = 2 * precision * recall / (precision + recall)
-    accuracy = (right_tag + 0.0) / all_tag
-    # print "Accuracy: ", right_tag,"/",all_tag,"=",accuracy
-    if label_type.upper().startswith("B"):
-        print("gold_num = ", golden_num, " pred_num = ", predict_num, " right_num = ", right_num)
-    else:
-        print("Right token = ", right_tag, " All token = ", all_tag, " acc = ", accuracy)
-    return accuracy, precision, recall, f_measure
-
-
-def reverse_style(input_string):
-    target_position = input_string.index('[')
-    input_len = len(input_string)
-    output_string = input_string[target_position:input_len] + input_string[0:target_position]
-    return output_string
-
-
-def get_ner_BMES(label_list):
-    # list_len = len(word_list)
-    # assert(list_len == len(label_list)), "word list size unmatch with label list"
-    list_len = len(label_list)
-    begin_label = 'B-'
-    end_label = 'E-'
-    single_label = 'S-'
-    whole_tag = ''
-    index_tag = ''
-    tag_list = []
-    stand_matrix = []
-    for i in range(0, list_len):
-        # wordlabel = word_list[i]
-        current_label = label_list[i].upper()
-        if begin_label in current_label:
-            if index_tag != '':
-                tag_list.append(whole_tag + ',' + str(i - 1))
-            whole_tag = current_label.replace(begin_label, "", 1) + '[' + str(i)
-            index_tag = current_label.replace(begin_label, "", 1)
-
-        elif single_label in current_label:
-            if index_tag != '':
-                tag_list.append(whole_tag + ',' + str(i - 1))
-            whole_tag = current_label.replace(single_label, "", 1) + '[' + str(i)
-            tag_list.append(whole_tag)
-            whole_tag = ""
-            index_tag = ""
-        elif end_label in current_label:
-            if index_tag != '':
-                tag_list.append(whole_tag + ',' + str(i))
-            whole_tag = ''
-            index_tag = ''
-        else:
-            continue
-    if (whole_tag != '') & (index_tag != ''):
-        tag_list.append(whole_tag)
-    tag_list_len = len(tag_list)
-
-    for i in range(0, tag_list_len):
-        if len(tag_list[i]) > 0:
-            tag_list[i] = tag_list[i] + ']'
-            insert_list = reverse_style(tag_list[i])
-            stand_matrix.append(insert_list)
-    # print stand_matrix
-    return stand_matrix
-
-
-def get_ner_BIO(label_list):
-    # list_len = len(word_list)
-    # assert(list_len == len(label_list)), "word list size unmatch with label list"
-    list_len = len(label_list)
-    begin_label = 'B-'
-    inside_label = 'I-'
-    whole_tag = ''
-    index_tag = ''
-    tag_list = []
-    stand_matrix = []
-    for i in range(0, list_len):
-        # wordlabel = word_list[i]
-        current_label = label_list[i].upper()
-        if begin_label in current_label:
-            if index_tag == '':
-                whole_tag = current_label.replace(begin_label, "", 1) + '[' + str(i)
-                index_tag = current_label.replace(begin_label, "", 1)
-            else:
-                tag_list.append(whole_tag + ',' + str(i - 1))
-                whole_tag = current_label.replace(begin_label, "", 1) + '[' + str(i)
-                index_tag = current_label.replace(begin_label, "", 1)
-
-        elif inside_label in current_label:
-            if current_label.replace(inside_label, "", 1) == index_tag:
-                whole_tag = whole_tag
-            else:
-                if (whole_tag != '') & (index_tag != ''):
-                    tag_list.append(whole_tag + ',' + str(i - 1))
-                whole_tag = ''
-                index_tag = ''
-        else:
-            if (whole_tag != '') & (index_tag != ''):
-                tag_list.append(whole_tag + ',' + str(i - 1))
-            whole_tag = ''
-            index_tag = ''
-
-    if (whole_tag != '') & (index_tag != ''):
-        tag_list.append(whole_tag)
-    tag_list_len = len(tag_list)
-
-    for i in range(0, tag_list_len):
-        if len(tag_list[i]) > 0:
-            tag_list[i] = tag_list[i] + ']'
-            insert_list = reverse_style(tag_list[i])
-            stand_matrix.append(insert_list)
-    return stand_matrix
