+++ src/requirements.txt
@@ -1,4 +1,3 @@
-apex==0.1
 boto3==1.9.130
 botocore==1.12.130
 certifi==2019.3.9
