--- src1/requirements.txt
@@ -1,3 +1,0 @@
-pytorch-pretrained-bert
-tqdm
-scipy
