--- src2/README.md
@@ -1,15 +1,0 @@
-# BERT Finetune
-
-### Install dependency
-```
-pip install -r requirements.txt
-```
-
-### Instructions
-
-- Train: Change and run the `./finetune.sh` script
-- Test: Change and run the `./predict.sh` script
-
-#### Arguments
-
-TODO: brief explanation for each argument
